Refactored the MiniMapRenderer to no longer be a singleton
It was made a singleton to be able to set the relevant map document to
be used by the MiniMapRenderer when used by the ResizeHelper. Instead,
the ResizeHelper now has a callback that can be set to customize how it
will draw the mini-map, so that it doesn't need to explicitly access the
MiniMapRenderer anymore.