Fixed fill tool for hexagonal maps with a staggered Y axis
This fixes the computation of the fill area for hexagonal maps using a
staggered Y axis (pointy-top hexagons). Maps using a staggered X axis and
isometric staggered maps still have problems and will be somewhat more
work to fix.

This kind of map was easy to fix because the hexes connect with each
other horizontally, which is a basic assumption made in the fill region
algorithm.

Issue #883