Implemented drag-n-drop support in Layers view
This makes it easy to move layers in and out of groups, as well as
changing their order.

The change required fixing the ReversingProxyModel to map drop
coordinates correctly and fixing ReparentLayers to deal with the layer
potentially not changing its parent but only its position within the
same parent.

Closes #1038