Fixed external JSON tilesets getting loaded multiple times
For external JSON tilesets, the check whether the tileset was already
loaded was missing. This caused duplicate tabs in the Tilesets view.

Solved this by handling this in the TilesetManager rather than in a
specialized MapReader, which only covered TSX tilesets. This is possible
now that the TilesetManager is part of libtiled.