Enable multi-property multi-object cut/copy/paste
Now you can cut/copy/paste multiple properties at the same time,
affecting all selected objects.

This turned out to be not so involved regarding the QtPropertyBrowser
solution, but all the more involved on the general behavior side. Care
had to be taken not to enable actions like cut/copy when it was not
clear what values would be taken or when the action could only be
applied to a subset of the selected property items.

Closes #515