Activate the Patreon dialog automatically after one month
When Tiled is used for at least a month and executed at least 7 times,
it will now notify the user that he could support the development of the
tool on Patreon.

The user can choose to suppress the reminder for some time or disable it
entirely, which should hopefully make this reminder acceptable.

Issue #1064