Fixed views changing size when toggling Clear View mode
Just hiding and then showing all views again seems to cause some layout
information to get lost or invalidated, which results in the views not
restoring to exactly the same sizes they had before.

For a toggle like the "Clear View" mode this is not acceptable so I've
changed the implementation to use saveState/restoreState instead. One
downside of this, is that any views made visible during the "Clear View"
mode will become hidden when the mode is untoggled, but this is less of
a problem that the views changing size.