Make sure tileset is added when inserting tile objects
This used to be a rare issue in Tiled 0.18 since the Tilesets view
normally didn't show tilesets that weren't part of the map. Starting
with Tiled 1.0, the Tilesets view shows all loaded tilesets (except
embedded tilesets in other maps) and the tool should really just make
sure to add the tileset when it isn't already there.

Closes #1571
Closes #1176