Fixed potential crash in Terrain Brush
The Terrain Brush could crash when you hold Shift while clicking outside
of the map, and then keep holding Shift while moving the mouse inside of
the map. This was due to missing out-of-bounds checks.

In addition, the tool could crash when the layer size didn't match the
map size. This scenario can't currently be achieved by using Tiled, but
the TMX format supports it and such a map could be crafted manually or
created with another editor.

Closes #1181