Allow selecting a folder to fix multiple broken links at once
When multiple broken links are selected, pop up a dialog to choose a
folder and search the chosen folder for files matching any of the broken
links.

A few minor changes remain to be done after the string freeze, like
adjusting some texts and adding better feedback, as well as cleaning up
the code.

Closes #1412