Windows: Work around an issue with OpenGL in full screen mode
When turning on OpenGL and using the full screen mode, menus and tooltips
were no longer visible. This is worked around by adding a one-pixel border
around the window.

See
http://doc.qt.io/qt-5/windows-issues.html#fullscreen-opengl-based-windows

Since this one-pixel border is annoying (it makes the top pixels of the
menu bar not clickable, for example), it is only enabled when necessary.

Closes #1576