Windows: Fixed bringing existing Tiled window to foreground
When Tiled is associated with certain file types and launched from the
Explorer (or generally, with files names in its parameters), it tries to
open these files in an existing Tiled instance when available.

On Windows, it failed to bring the window properly to the foreground, and
instead it would only be flashing on the task bar.

Closes #1256