Allow creating tilesets by dropping images on the Tilesets view again
This used to be possible in Tiled 0.18, but the functionality was lost
and then only the ability to drop external tilesets was added. Now it is
once again possible to create tilesets by dropping their image on the
Tilesets view, which pops up the New Tileset dialog for choosing the
parameters.

Thanks to Simon Park for pointing out the importance of keeping this
functionality
(http://discourse.mapeditor.org/t/drag-and-drop-tilesets/2378)