QtPropertyBrowser: Fixed reset of font size when using font dialog
Font dialogs generally deal with point sizes. When the font size is set
in pixels, convert to points before opening the dialog and then back to
pixels when the dialog is accepted.

One problem with this approach is that some font dialogs will round the
point size to the nearest integer, which means that simply opening and
accepting the dialog can cause the pixel size of the font to change. I
don't currently see a way around this.

Closes #1596