Revert the file format version back to "1.0"
This partially reverts ba7cffd12e61ec1a146141582a269b0a1fdb8ce4, which
wrote out the Tiled version as the version attribute in saved TMX and
JSON maps. However, the patch version should really never be relevant
and a full version like "1.0.0" is causing problems in some libraries
(like PyTmx), which parsed the version as a floating point number.

Additionally, using the Tiled version as file format version was a bad
idea, because the Tiled version may be something like "2017.06.12" for
snapshot builds, which would say very little about the format in which
it saves maps.

So, the TMX / JSON versions will now generally only change when there
were actual changes to the formats, and they will stay a two-part
number. A new attribute "tiledversion" was introduced for those who'd
like to know which version of Tiled wrote a particular file (similar to
how it was already done in the Lua export format).