Fixed hang when trying to fill with a pasted stamp
Since 688ec7d31c5efadb5be1525eb07772ebaa377e95 the size of a copied map
is set to 0x0 instead of matching the tile layer's size. It was supposed
to be irrelevant, but as it turns out TileStamp::maxSize was based on
the size of the map instead of the size of the tile layer. This could
lead to an infinite loop in fillWithStamp in bucketfilltool.cpp.

Closes #1617
Closes #1624