Implement read/write support for the tBIN format (#1622)
The tBIN format is used by the tile map editor tIDE and its rendering
engine xTile. This plugin allows Tiled to read and write tBIN files with
a high level of compatibility. 

The plugin is disabled by default, because the format does not support
all Tiled features. Using it unintentionally could lead to silent data
loss.

Supporting this format is interesting because it is used by Stardew 
Valley (issue #1560).