Changed tile, tileset and image layer images to use a URL
A URL is used in order to be able to refer to remote files. Custom
properties of type "file" now also support URLs.

Using a QUrl instead of a QString makes it easier to ensure all the
involved code parts are adjusted as needed.

This change does not enable actually loading remote files.