Limited the area processed by the Terrain Brush
This is mainly useful for making the brush work on infinite maps, but in
general it is good to prevent the tool from going crazy when it doesn't
have a complete set of transition tiles.

The area modified by the tool is now limitd to the maximum distance
between any two terrains, as counted by the number of transitions
needed.

A 'checked' boolean was added to each Cell so that the terrain tool can
use the chunked allocation logic of TileLayer, while still having a
fast place to store whether it processed a certain location or not.

This helps a lot with issue #1632