Added keyboard shortcut for Save All, Ctrl+Shift+L
There doesn't seem to be a strong standard for which shortcut to use. I
would prefer Ctrl+Shift+S because it's used in Visual Studio and Qt
Creator, and consistent with Ctrl+Shift+W for Close All.

However, Ctrl+Shift+S is already used as Save As on some platforms (e.g.
Gnome), which instead use Ctrl+Shift+L for Save All, so we follow suit.