Added keyboard shortcut for Save All, Ctrl+Shift+S
There doesn't seem to be a strong standard for which shortcut to use. I
chose Ctrl+Shift+S because it's used in Visual Studio and Qt Creator,
and consistent with Ctrl+Shift+W for Close All.

Some applications, like GEdit, use Ctrl+Shift+S for Save As instead, but
I think that's a much rarer operation which doesn't warrant the visual
clutter of a keyboard shortcut in the menu.