Added support for infinite maps (#1651)
This change adds initial support for infinite (or automatically 
expanding) maps. Such maps do not have a predefined size, and each tile
layer only allocates memory for the part that is actually used.

A lot of code needed to be adjusted to not rely on a fixed map or tile 
layer size. Of course, some new issues may still be hiding.

Known issues are:

* The storage format is based on the bounding rectangle of a tile 
  layer's content, which is sub-optimal for sparsely filled layers.
* Automapping does not currently work for infinite maps.

Closes #260