Fixed possible crash on undo/redo in collision editor
When there was a collision shape selected, the deselection of this
object on undo/redo could cause the mCurrentObject to be set to the
ObjectGroup instance right before it got deleted and stay around as
invalid pointer, which was accessed immediately afterwards when the
newly created ObjectGroup was set as current.

Solved by delaying the moment at which mCurrentObject is set to null
when layers get removed, which will hopefully not cause issues
elsewhere.

Closes #1695