Only write gid attribute for non-empty tiles
So in the (deprecated) pure XML tile layer data format, empty tiles are
now written as <tile/> instead of <tile gid="0"/>, reducing the file
size a little bit for largely empty layers.

This change works, because 0 is actually the default value of the gid
attribute (which is now also documented).

Closes #1716