Fixed issues caused by paths not being cleaned
Hopefully without introducing issues due to paths being cleaned...

Not cleaning the path here resulted in the same tile image being
referred to by "different" paths (like containing "foo/.." or "bar/..",
where "foo" and "bar" are both irrelevant due to the path containing
"..").

This caused for example the automapping to not recognize that two
embedded tilesets were the same, if the files they were part of were not
in the same directory.

Closes #1713