Windows: Fixed copy/paste pasting invalid tiles
Be more strict before considering a file path to be a URL. In particular,
never consider absolute file paths to be URLs.

This broke only on Windows, where the drive letters in absolute paths were
ending up as the "scheme" of the URL, making it non-relative. Problems
only appeared when copy/pasting, because in this case absolute paths are
used by the map in the clipboard.

Closes #1736