docs: Updated version and other minor changes
The version has been updated to 1.1.0, to match the upcoming Tiled
release. The manual already largely covers the new features in this
release.

Mentions of Tiled versions 0.10 and older have been removed, on the
grounds that this version is old enough that such information is no
longer useful. This excludes the TMX changelog of course.

Added a link to the GameMaker: Studio 1.4 export from the libraries and
frameworks page.