GmxPlugin: Added support for setting scale and origin for instances
The scale of instances was already determined automatically for tile
objects, but now it can also be manually specified or overridden using
the custom properties 'scaleX' and 'scaleY'.

It is now also possible to specify the origin of the object using custom
properties 'originX' and 'originY'. These will affect the position of the
exported instance, taking into account the rotation of the object as
well as its flipped state.

Closes #1427