GmxPlugin: Handle custom property inheritance
Now, object properties like scaleX, scaleY, originX, originY and also
all properties that can be set on "view" objects, are inherited for tile
objects, as well as for objects with a type (or have a type inherited
from their tile) and finally also from object templates.

This is essentially what you would expect, since those same values are
displayed in the Properties view.