Extend the scene rect as needed for infinite maps
This fixes the most serious issue with infinite maps, which was that it
was impossible to scroll towards newly painted area using the scroll
bars. It also fixes the size of the dark rectangle used to highlight the
current layer.

It does however introduce a new most serious issue, which is that the
updating of the scene rect is interfering with the tools due to it
causing a shift in the map view. There is also something wrong with the
undo for the lines that may be drawn as a result of the shift.