Fixed crash that could happen when painting outside current bounds
When painting outside of the current bounds, the bounds of an infinite
map will immediately increase. This currently causes the MapView to
immediately shift, which causes another mouse event that causes another
paint operation. That invalidated a local pointer.

To avoid such problems, even when we may later find a way to avoid the
shifting of the view, the local pointer is now a strong reference.