Fixed the view shifting when expanding an infinite map
This change improves the behavior of the "flexible scrollbars" (which
allow panning over the edge of the map) and changes the map view to
always include a margin, to avoid the forced centering of the map.

For infinite maps, this avoids the shifting of the view when expanding
the map, which interfered with the painting of tiles and caused problems
with the undo system due to generated mouse events happening while still
handling the actual mouse events.

I'm quite happy with the current behavior, and I think it makes the
zooming work well enough that it could be made the default mouse wheel
behavior (instead of requiring Ctrl).

Closes #1727