Added 'status' and 'wiki' properties
The 'wiki' property is expected to provide a link to the Tiled wiki
where more details about the given library are written. When set, it is
linked as "More information".

The 'status' property essentially replaced the 'updated' property, and
provides a simpler way of communicating the development activity of a
given library. The 'updated' property now indicates when this library
entry was last checked/updated.

Made the 'notes' property optional.