Moved most functionality of MapScene to the new MapItem
This graphics item should eventually enable the map view to show
multiple maps. For now this is just a refactoring, the functionality
should be exactly the same.

This change does also appear fix a repaint issue, by scheduling updates
in response to tile layer changes on the layer item instead of the
scene.