Fixed undo issues when layers and/or tilesets get added while painting
When a paint command gets merged into a previous paint command, it is
deleted along with its children. This is a problem now that we may merge
commands that have child commands for adding necessary layers and/or
tilesets.

Commands that may appear as children of commands that may be merged
should implement the new ClonableUndoCommand interface, which allows
them to be cloned as children of the previous command in case of
merging.