Automapping: Added a "StrictEmpty" flag to input layers
In "StrictEmpty" mode, empty tiles in the input region match empty tiles
in the set layer. So when an "input" layer contains an empty tile within
the input region, this means an empty tile is allowed at that location.
And when an "inputnot" layer contains an empty tile within the input
region, it means an empty tile is not allowed at that location.

Without the "StrictEmpty" option, when no "input" layer specifies any
tile for a given location, and no "inputnot" layers exist, any tile that
isn't used elsewhere on any input layer is considered valid. Also, when
an "inputnot" layer contains an empty tile, it is simply ignored. This
behavior can be convenient, but it is not always desired.

The "StrictEmpty" mode can be enabled per input layer by adding a custom
boolean property named "StrictEmpty" with the value "true".

See also the following discussion on the forum:
http://discourse.mapeditor.org/t/automapping-check-for-empty-tile/965