Tweaked code to avoid C++ compatibility warnings
This code worked with GCC, clang and even MSVC, but clang points out it
is relying on an extension to the standard. There is no real reason to
not just name the type.

warning: anonymous structs are a GNU extension
warning: anonymous types declared in an anonymous union are an extension