Fixed Object Selection tool clearing selection on double-click
Since the double-click event was not being handled by the map scene, the
tools only received the mouse release event. This caused the Object
Selection tool to clear the selection after a double-click.

Now the double-click event is handled and delivered to the active tool.
Currently, it's always routed to the press event, which fixes the
selection problem. In the future we could consider doing other on
double-click, like switching to the Edit Polygons tool when double-
clicking a polygon.