Qbs: Fixed linking on Linux and install Sparkle.framework
Setting 'cpp.frameworks' on Linux broke linking, so make sure to only use that
on macOS.

Also, now Sparkle.framework is properly installed into the Tiled.app. The
reason I previously was having problems with this was due to the automatically
applied file tags. Now the file tags are explicitly cleared to make sure the
files are only copied over. Thanks to @jakepetroules as always for his help.