Simplified code
There appears to be no difference between finishing or canceling the
extending of a polyline. In both cases we switch back to the "Edit
Polygons" tool and in both cases we should also make sure the object we
were extending is selected.

There shouldn't actually be a reason to touch the selection at all when
starting or stopping to extend a polyline, but currently the selection
outline does not update as expected while extending a polyline.