Allows extending of the polyline (#1683)
Extending can be done by right-clicking the nodes on either end of a polyline.

Implementation finished by @bjorn:
* Fixed new polylines being drawn backwards after extending first
* Fixed extending polyline other than the one just created
* Enabled switching between tools using the ToolManager
* Fixed behavior when switching tools while extending a polyline

The selection is cleared while extending a polyline. There shouldn't 
actually be a reason to touch the selection at all, but currently the
selection outline does not update as expected while extending a
polyline.

Issue #157.