Double-clicking on a polygon switches to Edit Polygons tool
Also double-clicking on polylines of course. This is a useful and fast
way of entering edit mode for polygons.

Pressed Escape in the "Edit Polygons" tool no longer clears the object
selection. It now first switches to the "Select Objects" tool, where
another press will clear the selection. I think this is a little more
useful and it works nicer in combination with this double-click
behavior.