JSON plugin: Fixed reading of infinite maps
In change fca34d6ad4d1216a31bdbf757749aa3505bca9e8 I forgot to also
adjust the VariantToMapConverter to handle chunked tile layer data,
probably because I based those changes on the earlier adaption of the
Lua format, which did not have a reader. This obviously causes problems
for those who use the JSON not only for exporting.

The JSON reader is now also capable of:

* Reading non-infinite maps that use the chunked data format.
* Reading infinite maps that do not use the chunked data format.

This may not be very useful right now since whether chunked data format
is used directly depends on whether the map is infinite or not, but in
the future this may change since the chunked format can also be
preferable for non-infinite maps.

Closes #1858