JSON plugin: Improved the JSON format (#1868)
The tiles in a tileset are now stored as an array instead of as an 
object. Previously the tile IDs were stored as string keys of the
"tiles" object, now they are stored as "id" property of each tile
object.

Similarly, the properties are now stored as an array instead of as an 
object where the property names were the keys. Each property in the
array is now an object that stores the name, type and value of the
property.

These changes should make the JSON format a lot easier to parse with 
certain JSON APIs. It avoids cases where the JSON format contained
objects that had no well defined set of members.

The version was bumped from 1.0 to 1.2.

Closes #1585