Fixed TileLayer::usedTilesets
Due to a scoping issue its behavior was rather strange. For an even
number of chunks, the returned tilesets would be the old set with
potentially some new tilesets added. For an odd number of chunks, none
of the tilesets used by even chunks would get counted.

This bug could lead to assertions, rendering issues and potentially
crashes. It was a regression introduced in fb8c380842915895c48549b0.