Tweaked logic of finding similar tilesets
If multiple similar tilesets exist in an automapping input or in a saved
tile stamp, replace all of them with the first one.

Previously, in this case the current map would have multiple similar
tilesets added to it, since for each of them it would only look for
similar tilesets within the set of existing ones.

This also unifies the logic of the unifyTilesets overrides after change
0b0921398989964ad43fb62c4941.