Objects view: Only center view on object on press or activation
Previously it would center the view on the selected objects as soon as
only one object was selected. This caused the centering to happen at
unwanted events, like when removing the last object from a selection or
when joining two selected polylines.

This removes the opening of the Properties view when pressing Enter on
an object selected in the Objects view, which now centers on the object
instead. I think this is not a bad change.