Fixed setting of changed flag when flipping objects
Related to issue #1895. However, it does not fix the whole problem,
since in fact a "gid" can't be written for template objects that use a
tile from a tileset that isn't also part of the map.

In addition, there is a caching issue causing tilesets used by both
templates and the map to get loaded twice. This is why it currently
won't even work for loaded maps, even if the tileset used by the
template is part of the map as well.