Removed manual reference counting of tilesets
Rely on the QSharedPointer for the lifetime and add/remove tilesets
instances from the TilesetManager directly from the Tileset class.

This fixes issues with tilesets getting loaded twice when they are used
by both a map and a template object, which happened as a result of the
manual adding of references being too late.

In general this approach should also be more reliable and prevent double
loading of tilesets in some cases.

Helps with issue #1895 when the tileset used by the template is also
part of the map.