Fixed exporting of external tilesets to JSON or TSX formats
When you tried to export an external tileset to JSON or TSX format, it
would export as a file with just a reference to the external tileset
file. Now it actually writes out the tileset like you would expect.

The File > Export As action only worked properly for the Lua format...