Python plugin: Fixed linking problems on Arch Linux
And probably fixed it on Ubuntu as well, with fallbacks to "python-3.6"
and "python-3.4".

The problem on Arch Linux was that the -lpython-3.6m linker flag was in
the wrong place in the arguments, causing the Python plugin to not
dynamically link to this library. This happens when directly passing
pkgConfigProbe.libs to cpp.linkerFlags.