Added snapcraft.yaml
This snapcraft.yaml builds a snap for Tiled 1.0.3. That's the best I
could do for the moment, because more recent versions of Tiled require a
more recent version of Qt (5.6) than is available on Ubuntu 16.04
(5.5.1).

Building a more recent version of Qt as a depencency is apparently
doable, but would increase the building time and maintenance for the
snap considerably. It is probably easier to patch up Tiled to build
against Qt 5.5.1 again.

Another open task is to use qbs instead of qmake to build the snap. A
plugin for this is already available so that shouldn't be a huge issue.

Finally, the Python plugin doesn't work for now, since I did not know
how to set it up so that it can find its libraries.