World: Added support for file name pattern matching
A pattern can define a regular expression, which needs to have two
captures to capture the X and Y values. It can then apply multipliers
and an offset to these values to derive the position of the map.

All files in the same directory as the .world that match against the
regular expression are considered part of that world.