libtiled: Introduced an ImageCache
This cache stores loaded images and the tiles that have been cut from
them, to avoid loading them again and again.

One problem with the cache currently is that when a tileset is loaded
and then closed, the image file is no longer monitored for changes. So
when the image is changed and then the tileset is opened again, the
outdated cached version will be used. To avoid this problem, a timestamp
could be saved.