Fixed shortcuts for flipping and rotating objects
This was an unintended regression in Tiled 1.1.0. Due to moving the
place where the shortcuts were defined, they only worked for tile stamps
and no longer worked for the Select Objects tool.

Now there flipping and rotating actions are also added to the tool bar
for all object-related tools.

The fix became rather involved, because I wanted to avoid doing virtual
function calls in the constructors of the tools.

A downside of setting the shortcuts on the menu bar actions is that they
don't work when the menu bar is hidden. The same is true for any such
actions at the moment. Maybe it makes sense to implement a generic
solution to this, as was done for the shortcuts to the tools themselves.

Closes #1926