Remove tile collision layer along with the last object
When removing the last collision shape, an empty "objectgroup" element
would stay behind in the saved file. Since this is not desired, it is
now removed when all collision objects are deleted.

I've experimented a little bit with using std::unique_ptr for
communicating pointer ownership in this change. It became rather more
verbose and I don't believe any bug is fixed, but the ownership
transfers are clearer now.

Closes #1230