Adjust Wang tile data when tileset column count changes
A change in the number of tile columns on a tileset image affects the
tile indices. This means that when loading a map that was saved before
the size of the tileset image was changed, or when reloading an image
while Tiled is open, we may need to adjust both map data and the
meta-data stored with the tileset.

This already happened for most of the data, but not for any Wang tile
related data. This change implements that and partly resolves
issue #1851.

It also fixes the test on whether to do the adjustment to also run when
just the tileset file is opened, as opposed to opening a map that uses
the tileset.

There is unfortunately still a problem when opening multiple maps that
need adjustment. In this case, the adjustment is only offered for the
first map. Fixing that will require adding more information to the
tileset references in each map. Alternatively we could think about
changing the tile indices entirely to make these adjustments no longer
necessary (for example, referring to the tiles by their x/y coordinate).