Fixed handling of sub-properties in Object Type Editor
For properties of type "color", sub-properties are displayed for each
color channel. When editing these, there were new top-level properties
being added with the name of the sub-property. They showed up in the
saved file and in the Object Type Editor after a restart.

This was due to the property value changed handler getting called also
for these sub-properties, and not ignoring them as it should.

Closes #1936