Avoid crash when deleting an object right before dragging it
When you start a drag operation by clicking an object, but press Delete
to delete it before you move the mouse, the application would crash due
to trying to use the deleted object.

This is now fixed by checking and clearing the roaming pointer to the
clicked object. The fix is not complete however, in that you could still
cause a roaming pointer by redoing the deletion of a layer that in turn
deletes the object. But that is rare and the fix is more complicated, so
I will fix that on master.

Closes #1933