Fixed memory management for Document instances
Replaced the DocumentRef class with a QSharedPointer<Document>, which is
the more convenient and less custom way of dealing with this. It did
require changing more existing code.

This change fixes a crash in the Tile Collision Editor, which was not
using MapDocumentRef and hence its dummy document was getting deleted
prematurely.