Clicking on a read-only map switches to an editable view for that map
This is a major usability improvement and an important part of the
multi-map view feature.

It may be that a modifier key like Control will be needed to avoid
clicking to another map too easily. Also, a solution will need to be
found for when maps have some overlap, since it is currently not
possible to interact with the part of the current map that has overlap
with a neighboring map.

Issue #1669