Python plugin: Fixed file change watching for main script files
Watch the actual script files in addition to the directory, which is
necessary for the reload to work when using certain editors, like
Sublime Text on Linux.

Note that this does not watch any script files in subdirectories that
the scripts may depend on, so in that case the reload does not work yet.