Templates view: Fixed a synchronization issue
Fixed an issue with synchronizating template edits when the same
template is opened in both the map editor and the tileset editor, by
sharing the dummy MapDocument instance that is created for editing the
template.

I think this is the first time that a MapDocument is shared between
multiple MapScene instances, but it seems to work fine.