Fixes to LayerIterator
Can't really remember how often I've been "fixing" up this tricky class.
Every time I think it's working now. Would be a perfect case for some
auto-tests...

This fixes a crash when trying to iterate a map without layers with a
range-based for loop, because the iterator could return a nullptr even
though it was unequal to the end() iterator.

It also fixes up LayerIterator::previous to keep going in case of
encountering layers that aren't of the expected type.