Fixed the background color of the Objects view
The custom viewport was not having the right background role and did not
have auto-fill background set, which caused the Objects view to show the
window background color.

The only reason for the custom viewport was to handle the leave event,
and I found that overriding viewportEvent provides a better way of doing
this.