Tile Collision Editor: Fixed potential crash on Undo
Undoing a change in the Tile Collision Editor causes the active tool to
get temporarily deactivated and the dummy MapDocument to be refreshed.
When the undo action was triggered while changing objects, the code
ended up referencing roaming pointers.

Now a bunch of variables are reset when the selection tool is
deactivated in order to avoid such crashes. A better fix will be to
properly reset the current action when the tool gets deactivated, but
since this code has changed in the meantime that will be implemented on
master.

Closes #1965