Object Selection Tool: Escape aborts current action
If you are currently performing some action, then Escape will now cancel
this action and bring the state back to when the action was started.

The current action is now also reset as appropriate when the tool gets
deactivated (for example by selecting to a different tool), or when some
objects get deleted (which still shouldn't happen, but it can be done
and it shouldn't crash).