Fixed crash when deleting the hovered object
That was a pretty nasty one to introduce, since it's quite common to
still hover the object you wanted to delete...

It should be enough to check whether the object is still part of an
ObjectGroup. The other reason the hovered object may be removed is when
its layer gets removed entirely. In that case it will still be part of
that layer, but since this is handled in "onLayerAboutToBeRemoved", it
happens before the object is actually removed.